Changed the UI controls to use databinding

+++ JomaxEmployeeDatabase/Services/DataService.cs
@@ -1,6 +1,8 @@
-﻿using JomaxEmployeeDatabase.Models;
+﻿using JomaxEmployeeDatabase.Helpers;
+using JomaxEmployeeDatabase.Models;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Data.SQLite;
 using System.Linq;
 using System.Text;
@@ -11,33 +13,40 @@ namespace JomaxEmployeeDatabase.Services
     public class DataService
     {
         private string connectionString = @"Data Source=K:\Users\smbid\source\repos\JomaxEmployeeDatabase\JomaxEmployeeDatabase\JomaxEmployeeDatabase.db3; Version=3; FailIfMissing=True; Foreign keys=True;";
-        private List<Employee> employees;
+        private SortableBindingList<Employee> employees;
         private List<Department> departments;
         private List<JobTitle> jobTitles;
-        private List<Certificate> certificates;
+        private SortableBindingList<Certificate> certificates;
 
         public DataService()
         {
-            employees = new List<Employee>();
+            employees = new SortableBindingList<Employee>();
             departments = new List<Department>();
             jobTitles = new List<JobTitle>();
-            certificates = new List<Certificate>();
+            certificates = new SortableBindingList<Certificate>();
         }
 
+        /// <summary>
+        /// Insert or update a record to the database.
+        /// </summary>
+        /// <param name="employee">Record to be inserted or updated.</param>
+        /// <returns>The record ID number associated with the inserted or updated record.</returns>
         public int UpdateEmployee(Employee employee)
         {
-            int result = -1;
+            long result = -1;
+
             using (SQLiteConnection conn = new SQLiteConnection(connectionString))
             {
                 conn.Open();
                 using (SQLiteCommand cmd = new SQLiteCommand(conn))
                 {
                     //If we are adding a new record
-                    if (employee.EmployeeID == 0)
+                    if (employee.Id == 0)
                     {
                         cmd.CommandText = "INSERT INTO employees " +
-                        "(firstName, lastName, jobTitle, department, VacationDaysInitial, VacationDaysAcrrued, VacationDaysUsed, DateCreated) VALUES " +
-                        "(@firstName, @lastName, @jobTitle, @department, @vacationDaysInitial, @vacationDaysAccrued, @vacationDaysUsed, @dateCreated)";
+                        "(firstName, lastName, jobTitle, department, VacationDaysInitial, VacationDaysAcrrued, VacationDaysUsed, DateCreated, dateModified, vacationComments) VALUES " +
+                        "(@firstName, @lastName, @jobTitle, @department, @vacationDaysInitial, @vacationDaysAccrued, @vacationDaysUsed, @dateCreated, @dateModified, @vacationComments) " +
+                        "returning employeeID";
                         employee.CreatedDate = DateTime.Now;
                     }
 
@@ -51,8 +60,11 @@ namespace JomaxEmployeeDatabase.Services
                             "department = @department, " +
                             "VacationDaysInitial = @vacationDaysInitial, " +
                             "VacationDaysAcrrued = @vacationDaysAccrued, " +
-                            "VacationDaysUsed = @vacationDaysUsed " +
-                            "WHERE employeeID = @employeeID";
+                            "VacationDaysUsed = @vacationDaysUsed, " +
+                            "dateModified = @dateModified, " +
+                            "vacationComments = @vacationComments " +
+                            "WHERE employeeID = @employeeID " +
+                            "returning employeeID";
                     }
                     cmd.Prepare();
                     cmd.Parameters.AddWithValue("@firstName", employee.FirstName);
@@ -62,63 +74,68 @@ namespace JomaxEmployeeDatabase.Services
                     cmd.Parameters.AddWithValue("@vacationDaysInitial", employee.VacationDaysInitial);
                     cmd.Parameters.AddWithValue("@vacationDaysAccrued", employee.VacationDaysAcrrued);
                     cmd.Parameters.AddWithValue("@vacationDaysUsed", employee.VacationDaysUsed);
-                    cmd.Parameters.AddWithValue("@employeeID", employee.EmployeeID);
+                    cmd.Parameters.AddWithValue("@employeeID", employee.Id);
                     cmd.Parameters.AddWithValue("@dateCreated", employee.CreatedDate);
+                    cmd.Parameters.AddWithValue("@dateModified", DateTime.Now);
+                    cmd.Parameters.AddWithValue("@vacationComments", employee.VacationComments);
 
                     try
                     {
-                        result=cmd.ExecuteNonQuery();
+                        result = (long)cmd.ExecuteScalar();
                     }
                     catch (SQLiteException ex)
                     {
                         MessageBox.Show(connectionString, ex.Message);
                     }
                 }
+                
                 conn.Close();
             }
-            return result;
+            return (int)result;
         }
 
         public int UpdateCertificate(Certificate certificate)
         {
-            int result = -1;
+            long result = -1;
             using (SQLiteConnection conn = new SQLiteConnection(connectionString))
             {
                 conn.Open();
                 using (SQLiteCommand cmd = new SQLiteCommand(conn))
                 {
                     //If we are adding a new record
-                    if (certificate.cerificationID == 0)
+                    if (certificate.Id == 0)
                     {
                         cmd.CommandText = "INSERT INTO certifications " +
                         "(employeeID, description, comments, dateInitial, dateExpires, doesExpire) VALUES " +
-                        "(@employeeID, @description, @comments, @dateInitial, @dateExpires, @doesExpire)";
+                        "(@employeeID, @description, @comments, @dateInitial, @dateExpires, @doesExpire) " +
+                        "returning certificationID";
                     }
 
                     //Update an existing record
                     else
                     {
-                        cmd.CommandText = "UPDATE employees " +
+                        cmd.CommandText = "UPDATE certifications " +
                             "SET employeeID = @employeeID, " +
                             "description = @description, " +
                             "comments = @comments, " +
                             "dateInitial = @dateInitial, " +
                             "dateExpires = @dateExpires, " +
-                            "doesExpire = @doesExpire " +                            
-                            "WHERE certificationID = @certificationID";
+                            "doesExpire = @doesExpire " +
+                            "WHERE certificationID = @certificationID " +
+                            "returning certificationID";
                     }
                     cmd.Prepare();
-                    cmd.Parameters.AddWithValue("@employeeID", certificate.employeeID);
-                    cmd.Parameters.AddWithValue("@description", certificate.description);
-                    cmd.Parameters.AddWithValue("@comments", certificate.comments);
-                    cmd.Parameters.AddWithValue("@dateInitial", certificate.dateInitial);
-                    cmd.Parameters.AddWithValue("@dateExpires", certificate.dateExpires);
-                    cmd.Parameters.AddWithValue("@doesExpire", certificate.doesExpire);
-                    cmd.Parameters.AddWithValue("@certificationID", certificate.cerificationID);
+                    cmd.Parameters.AddWithValue("@employeeID", certificate.EmployeeID);
+                    cmd.Parameters.AddWithValue("@description", certificate.Description);
+                    cmd.Parameters.AddWithValue("@comments", certificate.Comments);
+                    cmd.Parameters.AddWithValue("@dateInitial", certificate.DateInitial);
+                    cmd.Parameters.AddWithValue("@dateExpires", certificate.DateExpires);
+                    cmd.Parameters.AddWithValue("@doesExpire", certificate.DoesExpire);
+                    cmd.Parameters.AddWithValue("@certificationID", certificate.Id);
 
                     try
                     {
-                        result = cmd.ExecuteNonQuery();
+                        result = (long)cmd.ExecuteScalar();
                     }
                     catch (SQLiteException ex)
                     {
@@ -127,12 +144,12 @@ namespace JomaxEmployeeDatabase.Services
                 }
                 conn.Close();
             }
-            return result;
+            return (int)result;
         }
 
-        public IEnumerable<Employee> GetEmployees()
+        public SortableBindingList<Employee> GetEmployees()
         {
-            employees = new List<Employee>();
+            employees = new SortableBindingList<Employee>();
 
             try
             {
@@ -148,16 +165,17 @@ namespace JomaxEmployeeDatabase.Services
                         {
                             while (reader.Read())
                             {
-                                Employee employee = new Employee();
-                                employee.EmployeeID = (int)(long)reader["employeeID"];
-                                employee.FirstName = reader["firstName"].ToString() ?? "first name";
-                                employee.LastName = reader["lastName"].ToString() ?? "last name";
+                                Employee employee = Employee.CreateEmployee();
+                                employee.Id = (int)(long)reader["employeeID"];
+                                employee.FirstName = reader["firstName"].ToString() ?? String.Empty;
+                                employee.LastName = reader["lastName"].ToString() ?? String.Empty;
                                 employee.CreatedDate = DateTime.Parse((reader["DateCreated"].ToString() ?? DateTime.Now.ToString()));
-                                employee.JobTitle = reader["jobTitle"].ToString() ?? "Labourer";
-                                employee.Department = reader["department"].ToString() ?? "Shop";
+                                employee.JobTitle = reader["jobTitle"].ToString() ?? String.Empty;
+                                employee.Department = reader["department"].ToString() ?? String.Empty;
                                 employee.VacationDaysInitial = (int)(long)reader["vacationDaysInitial"];
                                 employee.VacationDaysAcrrued = (int)(long)reader["vacationDaysAcrrued"];
                                 employee.VacationDaysUsed = (int)(long)reader["vacationDaysUsed"];
+                                employee.VacationComments = reader["vacationComments"].ToString() ?? String.Empty;
 
                                 employees.Add(employee);
                             }
@@ -176,9 +194,9 @@ namespace JomaxEmployeeDatabase.Services
             return employees;
         }
 
-        public IEnumerable<Certificate> GetCertificates(int employeeID)
+        public SortableBindingList<Certificate> GetCertificates(int employeeID)
         {
-            certificates = new List<Certificate>();
+            certificates = new SortableBindingList<Certificate>();
 
             try
             {
@@ -194,16 +212,16 @@ namespace JomaxEmployeeDatabase.Services
                         {
                             while (reader.Read())
                             {
-                                Certificate certificate = new Certificate();
-                                certificate.cerificationID = (int)(long)reader["certificationID"];
-                                certificate.employeeID = (int)(long)reader["employeeID"];
-                                certificate.description = reader["description"].ToString();
-                                certificate.comments = reader["comments"].ToString();
-                                certificate.dateInitial = DateTime.Parse(reader["dateInitial"].ToString());                                
-                                certificate.doesExpire = Convert.ToBoolean((int)(long)reader["doesExpire"]);
-                                if (certificate.doesExpire)
+                                Certificate certificate = Certificate.CreateCertificate();
+                                certificate.Id = (int)(long)reader["certificationID"];
+                                certificate.EmployeeID = (int)(long)reader["employeeID"];
+                                certificate.Description = reader["description"].ToString() ?? String.Empty;
+                                certificate.Comments = reader["comments"].ToString() ?? String.Empty;
+                                certificate.DateInitial = DateTime.Parse((string)reader["dateInitial"]);
+                                certificate.DoesExpire = Convert.ToBoolean((int)(long)reader["doesExpire"]);
+                                if (certificate.DoesExpire)
                                 {
-                                    certificate.dateExpires = DateTime.Parse(reader["dateExpires"].ToString());
+                                    certificate.DateExpires = DateTime.Parse((string)reader["dateExpires"]);
                                 }
 
                                 certificates.Add(certificate);
@@ -223,19 +241,34 @@ namespace JomaxEmployeeDatabase.Services
             return certificates;
         }
 
-        public int DeleteEmployee(int id)
+        /// <summary>
+        /// Delete a record from the database
+        /// </summary>
+        /// <typeparam name="T">IDatabase object</typeparam>
+        /// <param name="dataObject">Object to be deleted from the database table</param>
+        /// <returns>The number of records affected</returns>
+        public int DeleteRecord<T>(IDatabaseObject dataObject) where T : IDatabaseObject
         {
             int result = -1;
-
+            T data = (T)dataObject;
+           
             using (SQLiteConnection conn=new SQLiteConnection(connectionString))
             {
                 conn.Open();
                 using (SQLiteCommand cmd = new SQLiteCommand(conn))
                 {
-                    cmd.CommandText = "DELETE FROM employees WHERE employeeID = @id";
+                    if (data.GetType() == typeof(Employee))
+                    {
+                        cmd.CommandText = "DELETE FROM employees WHERE employeeID = @id";
+                    }
+
+                    if (data.GetType() == typeof(Certificate))
+                    {
+                        cmd.CommandText = "DELETE FROM certifications WHERE certificationID = @id";
+                    }
 
                     cmd.Prepare();
-                    cmd.Parameters.AddWithValue("@id", id);
+                    cmd.Parameters.AddWithValue("@id", data.Id);
 
                     try
                     {
@@ -250,36 +283,7 @@ namespace JomaxEmployeeDatabase.Services
                 conn.Close();                
             }
             return result;
-        }
-
-        public int DeleteCertificate(int id)
-        {
-            int result = -1;
-
-            using (SQLiteConnection conn = new SQLiteConnection(connectionString))
-            {
-                conn.Open();
-                using (SQLiteCommand cmd = new SQLiteCommand(conn))
-                {
-                    cmd.CommandText = "DELETE FROM certifications WHERE certificationID = @id";
-
-                    cmd.Prepare();
-                    cmd.Parameters.AddWithValue("@id", id);
-
-                    try
-                    {
-                        result = cmd.ExecuteNonQuery();
-                    }
-
-                    catch (SQLiteException ex)
-                    {
-                        MessageBox.Show(ex.Message);
-                    }
-                }
-                conn.Close();
-            }
-            return result;
-        }
+        }        
 
         public IEnumerable<Department> GetDepartments()
         {
